Extract refresh logic to history graph view composable
